Alost done

+++ src/trash_file.rs
@@ -6,7 +6,7 @@ use std::{
     rc::Rc,
 };
 
-#[derive(Debug)]
+#[derive(Debug, Clone, PartialEq, Eq, PartialOrd, Ord, Hash)]
 pub struct TrashFile {
     trash: Rc<Trash>,
     trashinfo: TrashInfo,
@@ -105,12 +105,18 @@ impl TrashFile {
     /// directories of the file if they don't exist anymore.
     ///
     /// Returns the location the file was restored to
-    pub fn restore(self) -> crate::Result<PathBuf> {
+    pub fn restore(self, overwrite_existing: bool) -> crate::Result<PathBuf> {
         let original_path = self.original_path();
+        
+        if !overwrite_existing && fs::symlink_metadata(&original_path).is_ok() {
+            return Err(crate::Error::AlreadyExists(original_path));
+        }
+        
         if let Some(parent) = original_path.parent() {
             assert!(parent.is_absolute());
             fs::create_dir_all(parent)?;
         }
+
         fs::rename(self.files_filepath(), &original_path)?;
         fs::remove_file(self.info_filepath())?;
         Ok(original_path)
